requirements.txt
numpy~=2.1.2
scipy~=1.14.1
matplotlib~=3.9.2
setuptools~=70.0.0
pillow~=10.4.0